Move `mangled()` to the API module
